[processes] Fix build on win32, return error at run time
