+++ pkg/gohai/processes/processes.go
@@ -1,35 +1,15 @@
+// +build linux darwin
+
 package processes
 
 import (
-	"flag"
 	"strings"
 	"time"
 
 	"github.com/DataDog/gohai/processes/gops"
 )
 
-var options struct {
-	limit int
-}
-
 type ProcessField [7]interface{}
-
-type Processes struct{}
-
-const name = "processes"
-
-func init() {
-	flag.IntVar(&options.limit, name+"-limit", 20, "Number of process groups to return")
-}
-
-func (self *Processes) Name() string {
-	return name
-}
-
-func (self *Processes) Collect() (result interface{}, err error) {
-	result, err = getProcesses(options.limit)
-	return
-}
 
 // Return a JSON payload that's compatible with the legacy "processes" resource check
 func getProcesses(limit int) ([]interface{}, error) {
