+++ pkg/gohai/processes/processes_windows.go
@@ -1,0 +1,7 @@
+package processes
+
+import "errors"
+
+func getProcesses(limit int) ([]interface{}, error) {
+	return nil, errors.New("Not implemented on Windows")
+}
